Add files via upload
Added functionality for people to send me anonymous messages via email through an incorporation of Formspree's HTML to email service. Edited Formspree's suggested code so that only a message is required to be sent -- no longer does the form ask for a return address and no longer does the form require people to press a "Submit" button (the user is instead prompted to simply press the Enter key). I provided for the textbox to be transparent, larger, and flashing via learning some new CSS tricks.